# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OmniCreator AI Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/assets:
    get:
      tags: [Assets]
      operationId: listAssets
      summary: List assets in workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: folder
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: tags
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Asset list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Asset'
  /assets/{assetId}:
    get:
      tags: [Assets]
      operationId: getAsset
      summary: Get asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
    patch:
      tags: [Assets]
      operationId: updateAsset
      summary: Update asset metadata
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetBody'
      responses:
        '200':
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
    delete:
      tags: [Assets]
      operationId: deleteAsset
      summary: Delete asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /assets/upload-url:
    post:
      tags: [Assets]
      operationId: getUploadUrl
      summary: Get a signed upload URL for an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetUploadUrlBody'
      responses:
        '200':
          description: Upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'

  # AI JOBS
...
  /workspaces/{workspaceId}/analytics:
    get:
      tags: [Analytics]
      operationId: getWorkspaceAnalytics
      summary: Get workspace-level analytics
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
        - name: platform
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceAnalytics'
  /projects/{projectId}/analytics:
    get:
      tags: [Analytics]
      operationId: getProjectAnalytics
      summary: Get project analytics
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectAnalytics'
  /workspaces/{workspaceId}/analytics/recommendations:
    get:
      tags: [Analytics]
      operationId: getAnalyticsRecommendations
      summary: Get AI content recommendations
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalyticsRecommendation'

  # BILLING
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    # AUTH
...
    Asset:
      type: object
      required: [id, workspaceId, name, type, url, source, aiGenerated, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [image, video, audio, document, other]
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["number", "null"]
        width:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        duration:
          type: ["number", "null"]
        source:
          type: string
          enum: [uploaded, generated, stock, licensed]
        aiGenerated:
          type: boolean
        providerKey:
          type: ["string", "null"]
        promptId:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        folder:
          type: ["string", "null"]
        isFavorite:
          type: boolean
        commercialUseAllowed:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    UpdateAssetBody:
      type: object
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        folder:
          type: ["string", "null"]
        isFavorite:
          type: boolean
    GetUploadUrlBody:
      type: object
      required: [fileName, mimeType, workspaceId]
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        workspaceId:
          type: string
        folder:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadUrl, assetId, key]
      properties:
        uploadUrl:
          type: string
        assetId:
          type: string
        key:
          type: string

    # AI JOBS
...
    WorkspaceAnalytics:
      type: object
      required: [workspaceId, period]
      properties:
        workspaceId:
          type: string
        period:
          type: object
        projectsCreated:
          type: integer
        assetsGenerated:
          type: integer
        publishingSuccessRate:
          type: number
        totalCreditsUsed:
          type: number
        topPlatforms:
          type: array
          items:
            type: object
        postPerformance:
          type: array
          items:
            type: object
    ProjectAnalytics:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: string
        views:
          type: ["integer", "null"]
        likes:
          type: ["integer", "null"]
        comments:
          type: ["integer", "null"]
        shares:
          type: ["integer", "null"]
        watchTime:
          type: ["number", "null"]
        publishedPosts:
          type: array
          items:
            type: object
    AnalyticsRecommendation:
      type: object
      required: [id, type, title, reason, impact]
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        reason:
          type: string
        impact:
          type: string
        action:
          type: ["string", "null"]

    # BILLING
...