# packages/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OmniCreator AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetSource } from "./assetSource";
import type { AssetType } from "./assetType";

export interface Asset {
  id: string;
  workspaceId: string;
  name: string;
  type: AssetType;
  url: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  duration?: number | null;
  source: AssetSource;
  aiGenerated: boolean;
  /** @nullable */
  providerKey?: string | null;
  /** @nullable */
  promptId?: string | null;
  tags?: string[];
  /** @nullable */
  folder?: string | null;
  isFavorite?: boolean;
  /** @nullable */
  commercialUseAllowed?: boolean | null;
  createdAt: string;
  updatedAt: string;
}