# packages/api-zod/src/generated/types/projectAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OmniCreator AI Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectAnalyticsPublishedPostsItem } from "./projectAnalyticsPublishedPostsItem";

export interface ProjectAnalytics {
  projectId: string;
  /** @nullable */
  views?: number | null;
  /** @nullable */
  likes?: number | null;
  /** @nullable */
  comments?: number | null;
  /** @nullable */
  shares?: number | null;
  /** @nullable */
  watchTime?: number | null;
  publishedPosts?: ProjectAnalyticsPublishedPostsItem[];
}